// src/interfaces/appInterfaces.tsx
// Generated by https://quicktype.io

export interface LoginData {
  correo: string;
  password: string;
}

export interface RegisterData {
  nombre: string;
  correo: string;
  password: string;
}

export interface LoginResponse {
  usuario: User;
  token: string;
}

export interface User {
  rol: string;
  estado: boolean;
  google: boolean;
  nombre: string;
  correo: string;
  uid: string;
  img?: string;
}

// Products
// Generated by https://quicktype.io

export interface ProductsResponse {
  total: number;
  productos: Product[];
}

export interface Product {
  precio: number;
  _id: string;
  nombre: string;
  categoria: Category;
  usuario: Category;
  img?: string;
}

// Categories
// Generated by https://quicktype.io

export interface CategoriesResponse {
  total: number;
  categorias: Category[];
}

export interface Category {
  _id: string;
  nombre: string;
  usuario?: CreatedBy;
}

export interface CreatedBy {
  _id: string;
  nombre: string;
}